// src/app/core/interfaces/app/app-settings.interface.ts
export interface AppSettings {
    theme: string; // light, dark
    direction: string; // ltr, rtl
    layout: string; // vertical, horizontal
    sidebarType: string; // full, iconbar, overlay, mini-sidebar
    sidebarPosition: string; // fixed, absolute
    headerPosition: string; // fixed, absolute
    boxed: string; // full, boxed
    navbarBackground: string; // skin1, skin2, skin3, skin4, skin5, skin6
    sidebarBackground: string; // skin1, skin2, skin3, skin4, skin5, skin6
    logoBackground: string; // skin1, skin2, skin3, skin4, skin5, skin6
}